Split GPU ref fixture headers into infrastructure + shape catalogs
Separate shape data (withChannelLastLayout, get*Cases functions) from
test infrastructure (helpers, structs, fixture classes) so each concern
lives in its own header. This makes future tiering easier — new shapes
go into the catalog file without touching infrastructure.
