+++ dnn-providers/integration-tests/tests/gpu_ref/TestGpuFpReferenceDgrad.cpp
@@ -1,7 +1,7 @@
 // Copyright © Advanced Micro Devices, Inc., or its affiliates.
 // SPDX-License-Identifier:  MIT
 
-#include "GpuConvBwdRefTestFixture.hpp"
+#include "GpuConvBwdRefShapeCatalog.hpp"
 
 #include <cstdint>
 #include <stdexcept>
