+++ dnn-providers/integration-tests/tests/gpu_ref/TestGpuFpReferenceWgradSlow.cpp
@@ -1,7 +1,7 @@
 // Copyright © Advanced Micro Devices, Inc., or its affiliates.
 // SPDX-License-Identifier:  MIT
 
-#include "GpuConvWgradRefTestFixture.hpp"
+#include "GpuConvWgradRefShapeCatalog.hpp"
 
 using namespace gpu_conv_wgrad_ref_test;
